requirements.txt
pydantic>=2.0
google-genai>=1.68.0
numpy>=2.0
websockets>=12.0
python-dotenv>=1.0
